+++ crates/hyperqueue/src/dashboard/ui/screens/cluster_overview/overview/worker_count_chart.rs
