+++ crates/hyperqueue/src/dashboard/ui/screens/cluster_overview/worker/worker_config_table.rs
