+++ crates/hyperqueue/src/dashboard/ui/screens/mod.rs
@@ -1,4 +1,4 @@
 pub mod autoalloc_screen;
+pub mod cluster_overview;
 pub mod job_screen;
 pub mod root_screen;
-pub mod worker_overview_screen;
