--- crates/hyperqueue/src/dashboard/ui/screens/worker_overview_screen/overview/mod.rs
@@ -1,3 +1,0 @@
-pub mod fragment;
-pub mod worker_count_chart;
-pub mod worker_table;
